backend/assistant.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()
GROQ_API_KEY = os.getenv("GROQ_API_KEY")

def get_ai_response(prompt: str):
    url = "https://api.groq.com/openai/v1/chat/completions"
    headers = {
        "Authorization": f"Bearer {GROQ_API_KEY}",
        "Content-Type": "application/json"
    }
    payload = {
        "model": "llama3-8b-8192",
        "messages": [{"role": "user", "content": prompt}],
        "temperature": 0.7
    }

    response = requests.post(url, headers=headers, json=payload)

    # 👇 Check if the request was successful
    if response.status_code == 200:
        data = response.json()
        return {"response": data["choices"][0]["message"]["content"]}
    else:
        return {"error": f"Groq API error: {response.status_code}", "details": response.text}
